databases and catalogmethods

+++ platform/flowglad-next/seedDatabase.ts
@@ -216,6 +216,11 @@ export const teardownOrg = async ({
 }: {
   organizationId: string
 }) => {
+  if (process.env.VERCEL_ENV === 'production') {
+    throw new Error(
+      'attempted to access teardownOrg in production. This should never happen.'
+    )
+  }
   await sql`DELETE FROM "BillingPeriodItems" WHERE billingPeriodId IN (SELECT id FROM "BillingPeriods" WHERE subscriptionId IN (SELECT id FROM "Subscriptions" WHERE organizationId = ${organizationId}))`
   await sql`DELETE FROM "BillingRuns" WHERE billingPeriodId IN (SELECT id FROM "BillingPeriods" WHERE subscriptionId IN (SELECT id FROM "Subscriptions" WHERE organizationId = ${organizationId}))`
   await sql`DELETE FROM "Invoices" WHERE billingPeriodId IN (SELECT id FROM "BillingPeriods" WHERE subscriptionId IN (SELECT id FROM "Subscriptions" WHERE organizationId = ${organizationId}))`
